requirements.txt
torch
torchvision
tiktoken
webdataset
matplotlib
Pillow
webdataset